Feature/replace question

+++ app/src/main/java/com/example/onemillonwinner/ui/fragments/game/GameViewModel.kt
@@ -5,9 +5,11 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.example.onemillonwinner.data.GameQuestion
 import com.example.onemillonwinner.data.GameQuestionList
+import com.example.onemillonwinner.data.GameState
 import com.example.onemillonwinner.data.State
 import com.example.onemillonwinner.data.questionResponse.TriviaResponse
 import com.example.onemillonwinner.network.Repository
+import com.example.onemillonwinner.util.Constants.QUESTION_TIME
 import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
 import io.reactivex.rxjava3.core.Observable
 import io.reactivex.rxjava3.disposables.Disposable
@@ -18,39 +20,34 @@ class GameViewModel : ViewModel() {
 
     private val questionLogic: GameQuestionList by lazy { GameQuestionList() }
     private val repository: Repository by lazy { Repository() }
-
-    lateinit var timerDisposable: Disposable
+    private lateinit var timerDisposable: Disposable
 
     var isChangeQuestion = MutableLiveData(false)
     var isDeleteHalfOfAnswers = MutableLiveData(false)
     var isHelpByFriends = MutableLiveData(false)
 
-    private val _gameState = MutableLiveData<State<TriviaResponse>>()
-    val state: LiveData<State<TriviaResponse>>
+    private val _gameState = MutableLiveData<GameState>()
+    val state: LiveData<GameState>
         get() = _gameState
 
     private val _question = MutableLiveData<GameQuestion>()
     val question: LiveData<GameQuestion>
         get() = _question
 
-    private val _questionsLiveData = MutableLiveData<State<TriviaResponse>>()
-    val questions: LiveData<State<TriviaResponse>>
-        get() = _questionsLiveData
-
-    private val _questionTime = MutableLiveData(100)
+    private val _questionTime = MutableLiveData(QUESTION_TIME)
     val questionTime: LiveData<Int>
         get() = _questionTime
 
-    val questionTimeOver = MutableLiveData(false)
+    private val questionTimeOver = MutableLiveData(false)
 
 
     init {
-        _gameState.postValue(State.Loading)
+        _gameState.postValue(GameState.Loading)
         repository.getAllQuestions().subscribe(::onSuccessUpdateQuestion, ::onErrorUpdateQuestion)
     }
 
     private fun onSuccessUpdateQuestion(state: State<TriviaResponse>) {
-        _gameState.postValue(state)
+        _gameState.postValue(GameState.Success)
         timer()
         state.toData()?.let {
             questionLogic.setQuestions(it.questions)
@@ -59,54 +56,81 @@ class GameViewModel : ViewModel() {
     }
 
     private fun onErrorUpdateQuestion(throwable: Throwable) {
-
+        _gameState.postValue(GameState.Failure)
     }
 
     fun onClickToUpdateView() {
-        if (questionLogic.isReadyToSubmit()) {
-            questionLogic.setCurrentQuestionSubmitted(true)
-            _question.postValue(questionLogic.getCurrentQuestion())
-            timerDisposable.dispose()
-        } else if (questionLogic.isCurrentQuestionSubmitted()) {
-            questionLogic.setCurrentQuestionSubmitted(false)
-            updateView()
-        }
+       when(_gameState.value){
+           GameState.ANSWER_SELECTED->{
+               showAnswer()
+           }
+           GameState.QUESTION_SUBMITTED->{
+               updateView()
+           }
+           else -> {
+               _gameState.postValue(GameState.GameOver)
+           }
+       }
     }
+
+    private fun showAnswer() {
+        setGameState()
+        _question.postValue(questionLogic.getCurrentQuestion())
+        timerDisposable.dispose()
+    }
+
+    private fun setGameState() {
+        if (questionLogic.isGameOver())
+            _gameState.postValue(GameState.WRONG_ANSWER_SUBMITTED)
+        else
+            _gameState.postValue(GameState.QUESTION_SUBMITTED)
+    }
+
 
     private fun updateView() {
-        if (!questionLogic.isGameDone()) {
+        if (!questionLogic.isGameOver()) {
+            _gameState.postValue(GameState.QUESTION_START)
             _question.postValue(questionLogic.updateQuestion())
-            timerDisposable.dispose()
-            timer()
+            restartTimer()
         } else {
             timerDisposable.dispose()
-            _gameState.postValue(State.Complete)
-
+            _gameState.postValue(GameState.GameOver)
         }
     }
 
-    fun onClickAnswer(selectedAnswerIndex: Int) {
-        questionLogic.setSelectedAnswer(selectedAnswerIndex)
+    fun updateGameState(state: GameState) {
+        _gameState.postValue(state)
     }
 
-    fun changeQuestion() {
-        isChangeQuestion.postValue(true)
+    fun replaceQuestion() {
+        if (_gameState.value != GameState.QUESTION_SUBMITTED) {
+            _gameState.postValue(GameState.QUESTION_START)
+            isChangeQuestion.postValue(true)
+            _question.postValue(questionLogic.replaceQuestion())
+            restartTimer()
+        }
+    }
+
+    private fun restartTimer() {
+        timerDisposable.dispose()
+        timer()
     }
 
     fun deleteHalfOfAnswers() {
-        isDeleteHalfOfAnswers.postValue(true)
-        _question.postValue(questionLogic.deleteTwoWrongAnswersRandomly())
-
+        if (_gameState.value != GameState.QUESTION_SUBMITTED) {
+            _gameState.postValue(GameState.QUESTION_START)
+            isDeleteHalfOfAnswers.postValue(true)
+            _question.postValue(questionLogic.deleteTwoWrongAnswersRandomly())
+        }
     }
 
     fun helpByFriends() {
         isHelpByFriends.postValue(true)
     }
 
-
     private fun timer() {
-        _questionTime.postValue(100)
-        val timeInSecond: Long = 100
+        _questionTime.postValue(QUESTION_TIME)
+        val timeInSecond: Long = QUESTION_TIME.toLong()
         timerDisposable = Observable.interval(1, TimeUnit.SECONDS)
             .observeOn(AndroidSchedulers.mainThread())
             .take(timeInSecond).map {
@@ -121,9 +145,9 @@ class GameViewModel : ViewModel() {
 
     private fun endTheCountDown() {
         timerDisposable.dispose()
+        _gameState.postValue(GameState.GameOver)
         questionTimeOver.postValue(true)
     }
-
 
 }
